small correction
git-svn-id: https://192.168.1.150/svn/youtube_service@31 8123fb88-4fd9-be4e-917d-dc3803142b63

+++ app/models/Audiosettings.php
@@ -20,11 +20,13 @@ class AudioSettings extends Eloquent {
     public static function deleteEntry($id)
     {
         $audio = AudioSettings::find($id);
-        $audioName = public_path().'/audio/'.$audio->path;
-        if(file_exists($audioName)) {
-            unlink($audioName);
+        if(!empty($audio)) {
+            $audioName = public_path() . '/audio/' . $audio->path;
+            if (file_exists($audioName)) {
+                unlink($audioName);
+            }
+            $audio->delete();
         }
-        $audio->delete();
     }
 
     public static function getAudioByVideoId($videoId)
